+++ pallets/controller/src/lib.rs
@@ -38,6 +38,10 @@ pub use weights::WeightInfo;
 
 #[cfg(test)]
 mod tests;
+
+#[cfg(test)]
+mod mock;
+
 pub mod weights;
 
 #[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
@@ -848,7 +852,7 @@ impl<T: Config> Pallet<T> {
 	///
 	/// returns `utilization_rate = pool_borrow_underlying /
 	/// (pool_supply_underlying_balance + pool_borrow_underlying - pool_protocol_interest)`
-	pub fn calculate_utilization_rate(
+	fn calculate_utilization_rate(
 		pool_supply_underlying_balance: Balance,
 		pool_borrow_underlying: Balance,
 		pool_protocol_interest: Balance,
@@ -877,7 +881,7 @@ impl<T: Config> Pallet<T> {
 	/// - `accrual_block_number_previous`: Number of the last block with accruals.
 	///
 	/// returns `current_block_number - accrual_block_number_previous`
-	pub fn calculate_block_delta(
+	fn calculate_block_delta(
 		current_block_number: T::BlockNumber,
 		accrual_block_number_previous: T::BlockNumber,
 	) -> result::Result<T::BlockNumber, DispatchError> {
@@ -894,7 +898,7 @@ impl<T: Config> Pallet<T> {
 	/// - `block_delta`: The number of blocks elapsed since the last accrual.
 	///
 	/// returns `interest_factor = current_borrow_interest_rate * block_delta`.
-	pub fn calculate_interest_factor(current_borrow_interest_rate: Rate, block_delta: T::BlockNumber) -> RateResult {
+	fn calculate_interest_factor(current_borrow_interest_rate: Rate, block_delta: T::BlockNumber) -> RateResult {
 		let block_delta_as_usize = TryInto::<usize>::try_into(block_delta)
 			.ok()
 			.expect("blockchain will not exceed 2^32 blocks; qed");
