+++ pallets/dex/Cargo.toml
@@ -29,8 +29,6 @@ minterest-primitives = { path = "../../primitives" }
 liquidity-pools = { path = "../liquidity-pools" }
 liquidation-pools = { path = "../liquidation-pools" }
 test-helper = { path = "../../test-helper" }
-test-engine = { path = "../../test-engine" }
-dex = {path = "../dex"}
 
 [features]
 default = ["std"]
