+++ pallets/prices/src/tests.rs
@@ -1,19 +1,19 @@
 //! Unit tests for the prices module.
 
+use super::*;
+use crate::mock::{Event,*};
 use frame_support::{assert_noop, assert_ok};
 use minterest_primitives::Price;
-use module_prices::{Error, Event};
 use pallet_traits::PricesManager;
 use sp_runtime::{
 	traits::{BadOrigin, Zero},
 	FixedPointNumber,
 };
-use test_engine::*;
-pub use test_helper::*;
+
 
 #[test]
 fn get_underlying_price_should_work() {
-	ExtBuilderNew::default().build().execute_with(|| {
+	ExtBuilder::default().build().execute_with(|| {
 		// Price 1 BTC = 48000 USD
 		assert_eq!(
 			TestPrices::get_underlying_price(BTC),
@@ -33,7 +33,7 @@ fn get_underlying_price_should_work() {
 
 #[test]
 fn lock_price_should_work() {
-	ExtBuilderNew::default()
+	ExtBuilder::default()
 		.set_locked_price(BTC, Price::saturating_from_integer(80_000))
 		.build()
 		.execute_with(|| {
@@ -51,12 +51,12 @@ fn lock_price_should_work() {
 
 #[test]
 fn lock_price_call_should_work() {
-	ExtBuilderNew::default().build().execute_with(|| {
+	ExtBuilder::default().build().execute_with(|| {
 		System::set_block_number(1);
 		assert_ok!(TestPrices::lock_price(alice_origin(), BTC));
 
 		let lock_price_event =
-			test_engine::Event::TestPrices(Event::LockPrice(BTC, Price::saturating_from_integer(48000)));
+			Event::TestPrices(crate::Event::LockPrice(BTC, Price::saturating_from_integer(48000)));
 		assert!(System::events().iter().any(|record| record.event == lock_price_event));
 		assert_eq!(
 			TestPrices::locked_price_storage(BTC),
@@ -72,14 +72,14 @@ fn lock_price_call_should_work() {
 
 #[test]
 fn unlock_price_call_should_work() {
-	ExtBuilderNew::default()
+	ExtBuilder::default()
 		.set_locked_price(BTC, Price::saturating_from_integer(80000))
 		.build()
 		.execute_with(|| {
 			System::set_block_number(1);
 			assert_ok!(TestPrices::unlock_price(alice_origin(), BTC));
 
-			let unlock_price_event = test_engine::Event::TestPrices(Event::UnlockPrice(BTC));
+			let unlock_price_event = Event::TestPrices(crate::Event::UnlockPrice(BTC));
 			assert!(System::events().iter().any(|record| record.event == unlock_price_event));
 
 			assert_eq!(TestPrices::locked_price_storage(BTC), None);
