+++ test-helper/Cargo.toml
@@ -8,6 +8,12 @@ edition = "2018"
 [dependencies]
 minterest-primitives = { path = "../primitives" }
 frame-support = {  git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.8" }
+minterest-protocol = { path = "../pallets/minterest-protocol", default-features = false }
+liquidity-pools = { path = "../pallets/liquidity-pools", default-features = false }
+liquidation-pools = { path = "../pallets/liquidation-pools", default-features = false }
+minterest-model = { path = "../pallets/minterest-model", default-features = false }
+controller = { path = "../pallets/controller", default-features = false }
+
 
 [features]
 default = ["std"]
